refactor: update GameOptions import paths to use types module
